requirements.txt
aiogram==2.9.2
python-dotenv==0.14.0
emoji==0.6.0
morph==0.1.4